requirements.txt
moviepy
pandas
streamlit
flywheel
python-dotenv